Restructure and clean up code for separate transformations
See https://github.com/hbz/lobid-organisations/issues/41
