+++ app/controllers/Index.java
@@ -48,7 +48,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import play.Logger;
 import play.mvc.Controller;
 import play.mvc.Result;
-import transformation.Enrich;
+import transformation.TransformAll;
 
 /**
  * Indexing of the transformed data in elasticsearch
@@ -106,7 +106,7 @@ public class Index extends Controller {
 		if (!Application.CONFIG.getStringList("index.remote").contains(remote)) {
 			return forbidden();
 		}
-		initialize(Enrich.DATA_OUTPUT_FILE);
+		initialize(TransformAll.DATA_OUTPUT_FILE);
 		return ok("Started indexing");
 	}
