+++ app/controllers/Transformation.java
@@ -7,7 +7,7 @@ import java.io.IOException;
 import akka.event.slf4j.Logger;
 import play.mvc.Controller;
 import play.mvc.Result;
-import transformation.Enrich;
+import transformation.TransformAll;
 
 /**
  * Controller to start the transformation of the data
@@ -44,8 +44,8 @@ public class Transformation extends Controller {
 					Application.CONFIG.getString("transformation.updates.interval.size");
 			String geoLookupServer =
 					Application.CONFIG.getString("transformation.geo.lookup.server");
-			String outputPath = Enrich.DATA_OUTPUT_FILE;
-			Enrich.process(startOfUpdates, Integer.parseInt(intervalSize), outputPath,
+			String outputPath = TransformAll.DATA_OUTPUT_FILE;
+			TransformAll.process(startOfUpdates, Integer.parseInt(intervalSize), outputPath,
 					geoLookupServer);
 		} catch (Exception e) {
 			Logger.root().error("Transformation failed", e);
